requirements.txt
matplotlib==3.7.2
python-dotenv==1.0.0
Requests==2.31.0
